FindSIP: Respecting returncode
While looking at the problems with not found SIP installations on GitHub,
I found two solutions while looking at the code of FindSIP.cmake and FindSIP.py.

The Python script always prints output stdout, when the CMake script executes the Python script.
The CMake scripts checks whether stdout is empty or not. Currently it is always filled, once with the configs, on the other hand with the error message.
So we could make the Python script print to stderr instead.

I decided here to add a new variable, which catches the returncode, and finally made CMake check for the returncode except of the scripts output.
When looking at the Python script I guess this was the intention, as sys.exit(XY) was used in the script.

Fixes #41
Fixes #42