Update the build and install docs for improved clarity
The instructions mentioned using a cmake build directory in the middle
of discussing installing protobuf. Protobuf doesn't use cmake, so it was
likely a copy-paste error

Additionally there's not a great reason to fail to include 'sudo' in the
install command. Sudo is pretty wide-spread and it will trip people up
less to use a command they don't recognize than to expect them to
recognize the hash and then explain it after the fact.

Finally I removed the cmake-gui reference because it's not needed and
many people build on the console anyways. I couldn't find a cmake
variable to disable tests, if I had I would have added into the docs, so
I think this is a legacy issue that has gone away