Remove exception for modern CMake versions
The new FindPython3 script finds the wrong Python version. It finds the Python3.6 installed on the CI server from the OS, rather than the Python3.5.7 that we built manually. I'm going to try this as it seems to be the only difference between our build server (where it works) and our CI server (where it doesn't).