Allow Cmake to use FindPython3
This allows us to use FindPython3 if the cmake version
is higher or equal to 3.12. Which allows a user to set
`-DPython3_FIND_VIRTUALENV=ONLY` such that the library
can be installed in a virtual environment (mostly
usefull on Mac and Windows machines). This commit
doesn't break the current behaviour of our CI machines.