Fix getting PyCapsule error on import
This fix overcomes the notorious `ValueError: PyCapsule_GetPointer called with incorrect name` that
we get if we do not import Arcus, Savitar, and pynest2d in cura_app.py (and other places) even if it
is not used.

All credits go to Rex Dieter for figuring that this missing flag was the issue.
https://src.fedoraproject.org/rpms/libarcus/pull-request/1#request_diff

CURA-8146