Use the new CMake module add_sip_module
These changes are not fully tested locally. This is due to the way I
have personally setup my development environment.

This should generate a python module named `pyArcus.so` linking to in
Arcus.so both located in the build directory.

TODO: Add installing step.
TODO: propagate the name change to downstream dependencies

Contributes to CURA-7924