Hopefully, this'll resolve the issue where people force their own version of MSVCP140.dll instead of ours.
See https://github.com/actions/runner-images/issues/10004#issuecomment-2153445161 -- and that thread in general. Which is the same as an issue we _originally_ tried to work around by going with an older MSVC (which is what Lulzbot ended up going with) and when that turned out to not work, we just upgraded the entire chain to the new version. -- _However_ it now apprears that an antivirus vendor (McAfee) is forcing their own version of those DLL's, which are the older ones. -- So now try this workaround.

CURA-12675